Added read only file config provider

+++ core/src/main/java/net/lenni0451/optconfig/provider/impl/PathConfigProvider.java
@@ -10,10 +10,12 @@ public class PathConfigProvider implements ConfigProvider {
 
     private final Path path;
     private final Path tempPath;
+    private final boolean readOnly;
 
-    public PathConfigProvider(final Path path) {
+    public PathConfigProvider(final Path path, final boolean readOnly) {
         this.path = path;
         this.tempPath = path.resolveSibling(path.getFileName() + ".tmp");
+        this.readOnly = readOnly;
     }
 
     @Override
@@ -23,6 +25,7 @@ public class PathConfigProvider implements ConfigProvider {
 
     @Override
     public void save(byte[] content) throws IOException {
+        if (this.readOnly) return;
         if (this.path.getParent() != null) Files.createDirectories(this.path.getParent()); //Create parent directories if they don't exist
         Files.write(this.tempPath, content);
         Files.move(this.tempPath, this.path, java.nio.file.StandardCopyOption.REPLACE_EXISTING);
